requirements.txt
mcstatus==12.0.5
mcrcon==0.7.0
python-dotenv==1.0.0